backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Broker API integrations
ib_insync==0.9.86
# MetaTrader5==5.0.45  # Windows-only, using mock for development

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# WebSocket support
websockets==12.0

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Technical analysis
TA-Lib==0.4.28
ta==0.10.2

# Database (optional - for storing strategies, results)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver

# Caching and sessions
redis==5.0.1
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS support
fastapi-cors==0.0.6

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Financial libraries for quant trading
zipline-reloaded==3.0.4
backtrader==1.9.78.123
vectorbt==0.25.2
quantlib-python==1.32

# Machine learning (optional for advanced strategies)
scikit-learn==1.3.2
# tensorflow==2.15.0  # Uncomment if using ML strategies
# torch==2.1.1        # Uncomment if using PyTorch